# services/api/src/main/resources/META-INF/openapi-v1.yaml
openapi: 3.0.3
info:
  title: JavaBlog API
  description: REST API for javablog.com blog posts and comments
  version: 1.0.0
servers:
  - url: /v1
    description: API version 1
...
paths:
  /posts:
    post:
      tags:
        - Posts
      operationId: createPost
      summary: Create a new blog post
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Unauthorized - valid JWT required
    get:
      tags:
        - Posts
      parameters:
        - name: language
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/LanguageCode'
      operationId: listPosts
      summary: List all blog posts sorted by newest first
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostResponse'
  /posts/by-slug/{slug}:
    get:
      tags:
        - Posts
      operationId: getPostBySlug
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
  /posts/{postId}:
    get:
      tags:
        - Posts
      operationId: getPost
      summary: Get a blog post by ID
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
    put:
      tags:
        - Posts
      operationId: updatePost
      summary: Update a blog post
      security:
        - oidc: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Post not found
    delete:
      tags:
        - Posts
      operationId: deletePost
      summary: Delete a blog post
      security:
        - oidc: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Post deleted successfully
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Post not found
  /posts/{postId}/translations:
    get:
      tags:
        - Translations
      operationId: listTranslations
      summary: List translation state for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of translated posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslatedPost'
        '404':
          description: Post not found
    post:
      tags:
        - Translations
      operationId: requestTranslation
      summary: Request translation of a post into one or more languages
      security:
        - oidc: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationRequestBody'
      responses:
        '204':
          description: Translation request accepted
        '400':
          description: Invalid request - requested languages contain the original post language
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Post not found
  /posts/{postId}/comments:
    post:
      tags:
        - Comments
      operationId: createComment
      summary: Create a comment on a blog post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Post not found
    get:
      tags:
        - Comments
      operationId: listComments
      summary: List comments for a blog post sorted by newest first
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
        '404':
          description: Post not found
  /posts/{postId}/comments/{commentId}:
    delete:
      tags:
        - Comments
      operationId: deleteComment
      summary: Delete a comment
      security:
        - oidc: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: commentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Comment deleted successfully
        '401':
          description: Unauthorized - valid JWT required
        '404':
          description: Comment not found
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://keycloak.javablog.com/realms/javablog/.well-known/openid-configuration
      description: Keycloak OIDC authentication
  schemas:
    LanguageCode:
      type: string
      description: ISO 639-1 two-letter code (lowercase) for supported languages.
      enum:
        - bg   # Български
        - hr   # Hrvatski
        - cs   # Čeština
        - da   # Dansk
        - nl   # Nederlands
        - en   # English
        - et   # Eesti
        - fi   # Suomi
        - fr   # Français
        - de   # Deutsch
        - el   # Ελληνικά
        - hu   # Magyar
        - ga   # Gaeilge
        - it   # Italiano
        - lv   # Latviešu
        - lt   # Lietuvių
        - mt   # Malti
        - pl   # Polski
        - pt   # Português
        - ro   # Română
        - sk   # Slovenčina
        - sl   # Slovenščina
        - es   # Español
        - sv   # Svenska
        - ru   # Русский
        - uk   # Українська
        - be   # Беларуская
        - sq   # Shqip
        - "no" # Norsk
        - is   # Íslenska
        - ca   # Català
        - eu   # Euskara
        - gd   # Gàidhlig
        - cy   # Cymraeg
        - zh   # 简体中文
        - tw   # 臺灣華語
        - hi   # हिन्दी
        - ar   # العربية
        - id   # Bahasa Indonesia
        - bn   # বাংলা
        - ja   # 日本語
        - ko   # 한국어
        - tr   # Türkçe
        - he   # עברית
    TranslatedPost:
      type: object
      required:
        - orderedDatetime
        - language
      properties:
        orderedDatetime:
          type: string
          format: date-time
          description: When the translation was requested
        language:
          $ref: '#/components/schemas/LanguageCode'
        completedDatetime:
          type: string
          format: date-time
          description: When the translation was completed
        title:
          type: string
          description: Translated post title
        slug:
          type: string
          description: Translated post slug
    TranslationRequestBody:
      type: object
      required:
        - languages
      properties:
        languages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LanguageCode'
          description: Target languages for translation
    CreatePostRequest:
      type: object
      required:
        - slug
        - title
        - summary
        - content
        - language
      properties:
        slug:
          type: string
          description: URL-friendly identifier for the post
        title:
          type: string
          description: Post title
        summary:
          type: string
          maxLength: 500
          description: Short summary for listings
        content:
          type: string
          description: Post content
        language:
          $ref: '#/components/schemas/LanguageCode'
    UpdatePostRequest:
      type: object
      required:
        - slug
        - title
        - summary
        - content
        - language
      properties:
        slug:
          type: string
          description: URL-friendly identifier for the post
        title:
          type: string
          description: Post title
        summary:
          type: string
          maxLength: 500
          description: Short summary for listings
        content:
          type: string
          description: Post content
        language:
          $ref: '#/components/schemas/LanguageCode'
    CreateCommentRequest:
      type: object
      required:
        - author
        - content
      properties:
        author:
          type: string
          description: Comment author name
        content:
          type: string
          description: Comment content
    PostResponse:
      type: object
      required:
        - id
        - slug
        - title
        - summary
        - content
        - language
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique post identifier
        slug:
          type: string
          description: URL-friendly identifier
        title:
          type: string
          description: Post title
        summary:
          type: string
          description: Short summary for listings
        content:
          type: string
          description: Post content
        language:
          $ref: '#/components/schemas/LanguageCode'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
    CommentResponse:
      type: object
      required:
        - id
        - postId
        - author
        - content
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique comment identifier
        postId:
          type: string
          format: uuid
          description: Associated post identifier
        author:
          type: string
          description: Comment author name
        content:
          type: string
          description: Comment content
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
...